ai/data_collection/requirements.txt
bing-image-downloader==1.1.2
Pillow==11.1.0
torch==2.6.0
vision==1.0.0
transformers==4.40.1
